+++ packages/core/botpress/src/web/views/Dashboard/index.jsx
@@ -18,7 +18,8 @@ class Dashboard extends React.Component {
   state = {
     loading: true,
     popularModules: [],
-    featuredModules: []
+    featuredModules: [],
+    hiddenHeroSection: true
   }
 
   initialized = false
@@ -28,6 +29,8 @@ class Dashboard extends React.Component {
   }
 
   componentDidMount() {
+    this.fetchHeroConfig()
+
     this.refresh()
   }
 
@@ -44,7 +47,7 @@ class Dashboard extends React.Component {
   }
 
   refresh = () => {
-    if (this.initialized || !this.props.user || !this.props.user.id) {
+    if (this.initialized || !this.props.user || this.props.user.id == null) {
       return
     }
 
@@ -52,9 +55,15 @@ class Dashboard extends React.Component {
     this.queryAllModules().finally(() => this.setState({ loading: false }))
   }
 
+  fetchHeroConfig() {
+    axios
+      .get('/api/community/hero')
+      .then(({ data: { hidden: hiddenHeroSection } }) => this.setState({ hiddenHeroSection }))
+  }
+
   renderPopularModules() {
     return (
-      <PermissionsChecker user={this.props.user} op="read" res="modules.list.community">
+      <PermissionsChecker user={this.props.user} op="read" res="bot.modules.list.community">
         <Panel>
           <Panel.Heading>Popular modules</Panel.Heading>
           <Panel.Body>
@@ -67,7 +76,7 @@ class Dashboard extends React.Component {
 
   renderFeaturedModules() {
     return (
-      <PermissionsChecker user={this.props.user} op="read" res="modules.list.community">
+      <PermissionsChecker user={this.props.user} op="read" res="bot.modules.list.community">
         <Panel>
           <Panel.Heading>Featured modules</Panel.Heading>
           <Panel.Body>
@@ -78,10 +87,23 @@ class Dashboard extends React.Component {
     )
   }
 
+  hero() {
+    const { hiddenHeroSection } = this.state
+
+    return !hiddenHeroSection ? (
+      <Col xs={12} sm={8} md={4} smOffset={2} mdOffset={0}>
+        <HeroComponent />
+      </Col>
+    ) : null
+  }
+
   render() {
-    if (this.state.loading || !this.initialized) {
+    const { hiddenHeroSection, loading } = this.state
+
+    if (loading || !this.initialized) {
       return null
     }
+
     return (
       <ContentWrapper>
         <PageHeader>
@@ -89,14 +111,10 @@ class Dashboard extends React.Component {
         </PageHeader>
         <Grid fluid className={'bp-dashboard'}>
           <Row>
-            <Col sm={12} md={8}>
+            <Col sm={12} md={hiddenHeroSection ? 12 : 8}>
               <InformationComponent />
             </Col>
-            <Col xs={12} sm={8} md={4} smOffset={2} mdOffset={0}>
-              <PermissionsChecker user={this.props.user} res="modules.list.community" op="read">
-                <HeroComponent />
-              </PermissionsChecker>
-            </Col>
+            {this.hero()}
           </Row>
           <Row>
             <Col sm={12} md={6}>